gh-1381228: Split Modules/_remote_debugging_module.c into multiple files
