+++ PCbuild/_remote_debugging.vcxproj
@@ -98,7 +98,15 @@
   </Link>
 </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="..\Modules\_remote_debugging_module.c" />
+    <ClCompile Include="..\Modules\_remote_debugging\module.c" />
+    <ClCompile Include="..\Modules\_remote_debugging\object_reading.c" />
+    <ClCompile Include="..\Modules\_remote_debugging\code_objects.c" />
+    <ClCompile Include="..\Modules\_remote_debugging\frames.c" />
+    <ClCompile Include="..\Modules\_remote_debugging\threads.c" />
+    <ClCompile Include="..\Modules\_remote_debugging\asyncio.c" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="..\Modules\_remote_debugging\_remote_debugging.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\PC\python_nt.rc" />
